requirements.txt
torch==1.12.1  # PyTorch
torchvision==0.13.1  # For image transformations and model architectures
numpy==1.21.6  # For numerical operations
Pillow==9.1.0  # For image handling
scikit-learn==1.0.2  # For machine learning utilities
matplotlib==3.5.1  # For plotting
pandas==1.3.5  # For data manipulation (if used)
